+++ Dominos/bars/actionBar/bindings.lua
@@ -56,9 +56,9 @@ Binder:SetScript("OnEvent", function(self, event)
     self:SetScript("OnEvent", nil)
     self:UnregisterEvent(event)
 
-    self:SetAttribute("overrideui", OverrideActionBarButton1:IsShown() and 1 or 0)
+    self:SetAttribute("overrideui", OverrideActionBarButton1:IsVisible() and 1 or 0)
     self:SetAttribute("useoverrideui", Addon:UsingOverrideUI() and 1 or 0)
-    RegisterAttributeDriver(self, "petabattleui", "[petbattle]1;0")
+    RegisterAttributeDriver(self, "petbattleui", "[petbattle]1;0")
 end)
 
 Binder:RegisterEvent("PLAYER_LOGIN")