+++ Dominos/bars/actionBar/buttons.lua
@@ -21,10 +21,6 @@ ActionButtons.ShowGridReasons = {
     SHOW_EMPTY_BUTTONS_PER_BAR = 64
 }
 
--- how many bars are available
-local function IsSummonPetAction(action)
-    return GetActionInfo(action) == "summonpet"
-end
 
 local function HasSpellID(action, spellID)
     local actionType, id, subType = GetActionInfo(action)
@@ -43,20 +39,19 @@ local function HasSpellID(action, spellID)
     return false
 end
 
+local function SafeMixin(button, trait)
+    for k, v in pairs(trait) do
+        if rawget(button, k) ~= nil then
+            error(("%s[%q] has alrady been set"):format(button:GetName(), k), 2)
+        end
+
+        button[k] = v
+    end
+end
+
 -- states
 -- [button] = action
 ActionButtons.buttons = {}
-
--- [action] = { [button] = true }
-ActionButtons.actionButtons = setmetatable({}, {
-    __index = function(t, k)
-        local r = {}
-
-        t[k] = r
-
-        return r
-    end
-})
 
 -- dirty secure attributes
 ActionButtons.dirtyCvars = {}
@@ -78,46 +73,15 @@ function ActionButtons:PLAYER_LOGIN()
     -- initialize state
     self:SetAttributeNoHandler("showgrid", 0)
     self:SetAttribute("lockActionBars", GetCVarBool("lockActionBars"))
-    self:SetShowSpellGlows(Addon:ShowingSpellGlows())
-    self:SetShowGrid(Addon:ShowGrid(), self.ShowGridReasons.SHOW_EMPTY_BUTTONS)
 
     -- game events
-    self:TryRegisterEvent("ACTION_RANGE_CHECK_UPDATE")
-    self:TryRegisterEvent("ACTION_USABLE_CHANGED")
     self:TryRegisterEvent("ACTIONBAR_HIDEGRID")
     self:TryRegisterEvent("ACTIONBAR_SHOWGRID")
-    self:TryRegisterEvent("ACTIONBAR_SLOT_CHANGED")
-    self:TryRegisterEvent("ACTIONBAR_UPDATE_COOLDOWN")
-    self:TryRegisterEvent("ACTIONBAR_UPDATE_STATE")
-    self:TryRegisterEvent("ARCHAEOLOGY_CLOSED")
-    self:TryRegisterEvent("COMPANION_UPDATE")
-    self:TryRegisterEvent("CVAR_UPDATE")
-    self:TryRegisterEvent("PET_BAR_UPDATE")
-    self:TryRegisterEvent("PET_STABLE_SHOW")
-    self:TryRegisterEvent("PET_STABLE_UPDATE")
-    self:TryRegisterEvent("PLAYER_ENTER_COMBAT")
     self:TryRegisterEvent("PLAYER_ENTERING_WORLD")
-    self:TryRegisterEvent("PLAYER_LEAVE_COMBAT")
-    self:TryRegisterEvent("PLAYER_REGEN_ENABLED")
-    self:TryRegisterEvent("SPELL_UPDATE_CHARGES")
-    self:TryRegisterEvent("SPELL_UPDATE_ICON")
-    self:TryRegisterEvent("START_AUTOREPEAT_SPELL")
-    self:TryRegisterEvent("STOP_AUTOREPEAT_SPELL")
-    self:TryRegisterEvent("TRADE_SKILL_CLOSE")
-    self:TryRegisterEvent("TRADE_SKILL_SHOW")
-    self:TryRegisterEvent("UPDATE_SHAPESHIFT_FORM")
-    self:TryRegisterEvent("UPDATE_SUMMONPETS_ACTION")
-
-    -- unit events
-    self:TryRegisterUnitEvent("LOSS_OF_CONTROL_ADDED", "player")
-    self:TryRegisterUnitEvent("LOSS_OF_CONTROL_UPDATE", "player")
-    self:TryRegisterUnitEvent("UNIT_AURA", "pet")
-    self:TryRegisterUnitEvent("UNIT_ENTERED_VEHICLE", "player")
-    self:TryRegisterUnitEvent("UNIT_EXITED_VEHICLE", "player")
-    self:TryRegisterUnitEvent("UNIT_FLAGS", "pet")
 
     -- addon callbacks
     Addon.RegisterCallback(self, "SHOW_EMPTY_BUTTONS_CHANGED")
+    Addon.RegisterCallback(self, "SHOW_SPELL_ANIMATIONS_CHANGED")
     Addon.RegisterCallback(self, "SHOW_SPELL_GLOWS_CHANGED")
     Addon.RegisterCallback(self, "LAYOUT_LOADED")
 
@@ -172,33 +136,6 @@ function ActionButtons:PLAYER_LOGIN()
     ]])
 end
 
-function ActionButtons:ACTION_RANGE_CHECK_UPDATE(slot, isInRange, checksRange)
-    local buttons = self.actionButtons[slot]
-
-    if buttons then
-        local usable, oom = IsUsableAction(slot)
-        local oor = checksRange and not isInRange
-
-        for button in pairs(buttons) do
-            button:UpdateUsable(usable, oom, oor)
-        end
-    end
-end
-
-function ActionButtons:ACTION_USABLE_CHANGED(changes)
-    for _, change in pairs(changes) do
-        local buttons = self.actionButtons[change.slot]
-
-        if buttons ~= nil then
-            local oor = IsActionInRange(change.slot) == false
-
-            for button in pairs(buttons) do
-                button:UpdateUsable(change.usable, change.noMana, oor)
-            end
-        end
-    end
-end
-
 function ActionButtons:ACTIONBAR_SHOWGRID()
     self:SetShowGrid(true, self.ShowGridReasons.GAME_EVENT)
 end
@@ -207,60 +144,8 @@ function ActionButtons:ACTIONBAR_HIDEGRID()
     self:SetShowGrid(false, self.ShowGridReasons.GAME_EVENT)
 end
 
-function ActionButtons:ACTIONBAR_UPDATE_STATE()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:ACTIONBAR_SLOT_CHANGED(slot)
-    if slot == 0 or slot == nil then
-        self:ForAll("Update")
-    else
-        self:ForActionSlot(slot, "Update")
-    end
-end
-
-function ActionButtons:ACTIONBAR_UPDATE_COOLDOWN()
-    self:ForAllWhere(HasAction, "UpdateCooldown")
-end
-
-function ActionButtons:ARCHAEOLOGY_CLOSED()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:CVAR_UPDATE(name, ...)
-    if name == "lockActionBars" then
-        self:TrySetCVarAttribute(name, GetCVarBool(name))
-    end
-end
-
-function ActionButtons:COMPANION_UPDATE(companionType)
-    if companionType == "MOUNT" then
-        self:ForAllWhere(HasAction, "UpdateActive")
-    end
-end
-
-function ActionButtons:PET_BAR_UPDATE()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:PET_STABLE_SHOW()
-    self:ForAll("Update")
-end
-
-function ActionButtons:PET_STABLE_UPDATE()
-    self:ForAll("Update")
-end
-
-function ActionButtons:PLAYER_ENTER_COMBAT()
-    self:ForAllWhere(IsAttackAction, "UpdateFlashing")
-end
-
-function ActionButtons:PLAYER_LEAVE_COMBAT()
-    self:ForAllWhere(IsAttackAction, "UpdateFlashing")
-end
-
 function ActionButtons:PLAYER_ENTERING_WORLD()
-    self:ForAll("Update")
+    self:ForAll("UpdateShown")
 end
 
 function ActionButtons:PLAYER_REGEN_ENABLED()
@@ -268,71 +153,6 @@ function ActionButtons:PLAYER_REGEN_ENABLED()
         self:SetAttribute(k, GetCVarBool(k))
         self.dirtyCvars[k] = nil
     end
-end
-
-function ActionButtons:SPELL_UPDATE_CHARGES()
-    self:ForAllWhere(HasAction, "UpdateCount")
-end
-
-function ActionButtons:SPELL_ACTIVATION_OVERLAY_GLOW_SHOW(spellID)
-    self:ForSpellID(spellID, "ShowOverlayGlow")
-end
-
-function ActionButtons:SPELL_ACTIVATION_OVERLAY_GLOW_HIDE(spellID)
-    self:ForSpellID(spellID, "HideOverlayGlow")
-end
-
-function ActionButtons:SPELL_UPDATE_ICON()
-    self:ForAllWhere(HasAction, "UpdateIcon")
-end
-
-function ActionButtons:START_AUTOREPEAT_SPELL()
-    self:ForAllWhere(IsAutoRepeatAction, "UpdateFlashing")
-end
-
-function ActionButtons:STOP_AUTOREPEAT_SPELL()
-    self:ForAllWhere(IsAutoRepeatAction, "UpdateFlashing")
-end
-
-function ActionButtons:TRADE_SKILL_CLOSE()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:TRADE_SKILL_SHOW()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:UPDATE_SHAPESHIFT_FORM()
-    self:ForAllWhere(HasAction, "UpdateIcon")
-end
-
-function ActionButtons:UPDATE_SUMMONPETS_ACTION()
-    self:ForAllWhere(IsSummonPetAction, "UpdateIcon")
-end
-
--- unit events
-function ActionButtons:LOSS_OF_CONTROL_ADDED()
-    self:ForAllWhere(HasAction, "UpdateCooldown")
-end
-
-function ActionButtons:LOSS_OF_CONTROL_UPDATE()
-    self:ForAllWhere(HasAction, "UpdateCooldown")
-end
-
-function ActionButtons:UNIT_AURA()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:UNIT_ENTERED_VEHICLE()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:UNIT_EXITED_VEHICLE()
-    self:ForAllWhere(HasAction, "UpdateActive")
-end
-
-function ActionButtons:UNIT_FLAGS()
-    self:ForAllWhere(HasAction, "UpdateActive")
 end
 
 -- addon callbacks
@@ -348,29 +168,25 @@ function ActionButtons:SHOW_EMPTY_BUTTONS_CHANGED(_, show)
     self:SetShowGrid(show, self.ShowGridReasons.SHOW_EMPTY_BUTTONS)
 end
 
+function ActionButtons:SHOW_SPELL_ANIMATIONS_CHANGED(_, show)
+    self:SetShowSpellAnimations(show)
+end
+
 function ActionButtons:SHOW_SPELL_GLOWS_CHANGED(_, show)
     self:SetShowSpellGlows(show)
 end
 
 function ActionButtons:LAYOUT_LOADED()
     self:SetShowSpellGlows(Addon:ShowingSpellGlows())
+    self:SetShowSpellAnimations(Addon:ShowingSpellAnimations())
     self:SetShowGrid(Addon:ShowGrid(), self.ShowGridReasons.SHOW_EMPTY_BUTTONS)
 end
 
-function ActionButtons:OnActionChanged(buttonName, action, prevAction)
+function ActionButtons:OnActionChanged(buttonName, action)
     local button = _G[buttonName]
-    if button == nil then
-        return
+    if button ~= nil then
+        self.buttons[button] = action
     end
-
-    if prevAction ~= nil then
-        self.actionButtons[prevAction][button] = nil
-    end
-
-    self.actionButtons[action][button] = action
-    self.buttons[button] = action
-
-    C_ActionBar.EnableActionRangeCheck(action, true)
 end
 
 --------------------------------------------------------------------------------
@@ -383,75 +199,15 @@ local ActionButton_AttributeChanged = [[
     local prevValue = ActionButtons[self]
     if prevValue ~= value then
         ActionButtons[self] = value
+        self:RunAttribute("UpdateShown")
         control:CallMethod("OnActionChanged", self:GetName(), value, prevValue)
     end
 ]]
 
--- pre click:
--- update press and hold action state
-local ActionButton_PreClick = [[
-    local actionType, id = GetActionInfo(self:GetAttribute("action"))
-
-    if actionType == "spell" then
-        local ph = IsPressHoldReleaseSpell(id)
-        if self:GetAttribute("pressAndHoldAction") ~= ph then
-            self:SetAttribute("pressAndHoldAction", ph)
-        end
-    end
-]]
-
--- on click:
--- remap hotkey presses to LeftButton and let both down and up clicks through
--- prevent activating actions on mouse button clicks. This is to avoid conflicts
--- with drag and drop behaviors
---
--- When filtering out mouse button down presses, we need to also temporarily
--- turn off the cast on key down behavior. We restore it after the mouse button
--- was released
---
--- /click macros complicate this a bit. The simplest version, /click Button only
--- triggers the default click (left button up). So to handle these, we keep
--- track of the button that was originally clicked. The original button will
--- disable the cast on key press setting for any mouse button down call.
--- Any buttons clicked during the click of that button will adjust the setting
--- on an up click
 local ActionButton_Click = [[
-    local callerName = control:GetAttribute("caller")
-    local buttonName = self:GetName()
-    local isCaller
-
-    if callerName == nil then
-        control:SetAttribute("caller", buttonName)
-        isCaller = true
-    else
-        isCaller = callerName == buttonName
-    end
-
     if button == "HOTKEY" then
         return "LeftButton"
     end
-
-    if down then
-        if isCaller then
-            control:CallMethod("SaveActionButtonUseKeyDown", buttonName)
-        end
-        return false
-    end
-
-    if not isCaller then
-        control:CallMethod("SaveActionButtonUseKeyDown", buttonName)
-    end
-    return nil, true
-]]
-
-local ActionButton_ClickAfter = [[
-    local buttonName = self:GetName()
-
-    if control:GetFrameRef("caller") == buttonName then
-        control:SetFrameRef("caller", nil)
-    end
-
-    control:CallMethod("RestoreActionButtonUseKeyDown", buttonName)
 ]]
 
 -- post click:
@@ -460,13 +216,6 @@ local ActionButton_ClickAfter = [[
 -- mouse button, and then clicked on the button to place an action
 local ActionButton_PostClick = [[
     control:RunAttribute("ForActionSlot", self:GetAttribute("action"), "UpdateShown")
-]]
-
--- drag & drop
-local ActionButton_DragStart = [[
-    if not (IsModifiedClick("PICKUPACTION") or not control:GetAttribute("lockActionBars")) then
-        return false
-    end
 ]]
 
 local ActionButton_ReceiveDragBefore = [[
@@ -488,19 +237,15 @@ function ActionButtons:GetOrCreateActionButton(id, parent)
     local button = _G[name]
 
     if button == nil then
-        button = CreateFrame("CheckButton", name, parent, "SecureActionButtonTemplate, SecureHandlerAttributeTemplate, SecureHandlerDragTemplate, ActionButtonTemplate")
+        button = CreateFrame("CheckButton", name, parent, "ActionBarButtonTemplate")
 
-        Addon.ActionButton:Bind(button)
+        SafeMixin(button, Addon.ActionButton)
 
         button:OnCreate(id)
 
         self:WrapScript(button, "OnAttributeChanged", ActionButton_AttributeChanged)
-
-        self:WrapScript(button, "PreClick", ActionButton_PreClick)
-        self:WrapScript(button, "OnClick", ActionButton_Click, ActionButton_ClickAfter)
+        self:WrapScript(button, "OnClick", ActionButton_Click)
         self:WrapScript(button, "PostClick", ActionButton_PostClick)
-
-        self:WrapScript(button, "OnDragStart", ActionButton_DragStart)
         self:WrapScript(button, "OnReceiveDrag", ActionButton_ReceiveDragBefore, ActionButton_ReceiveDragAfter)
 
         -- register the button with the controller
@@ -516,35 +261,59 @@ function ActionButtons:GetOrCreateActionButton(id, parent)
 end
 
 function ActionButtons:SetShowGrid(show, reason)
-    self:ForAll("SetShowGrid", show, reason)
+    self:ForAll("SetShowGridInsecure", show, reason)
 end
 
 function ActionButtons:SetShowSpellGlows(enable)
+    local f = ActionBarActionEventsFrame
+
     if enable then
-        if not self:IsEventRegistered("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW") then
-            self:TryRegisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_HIDE")
-            self:TryRegisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW")
+        if not f:IsEventRegistered("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW") then
+            f:RegisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW")
+            f:RegisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_HIDE")
         end
     else
-        if self:IsEventRegistered("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW") then
-            self:UnregisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_HIDE")
-            self:UnregisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW")
-            self:ForAll("HideOverlayGlow")
+        if f:IsEventRegistered("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW") then
+            f:UnregisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_SHOW")
+            f:UnregisterEvent("SPELL_ACTIVATION_OVERLAY_GLOW_HIDE")
         end
     end
 end
 
-function ActionButtons:SaveActionButtonUseKeyDown(owner)
-    if self.restoreKeyDown == nil and GetCVarBool("ActionButtonUseKeyDown") then
-        SetCVar("ActionButtonUseKeyDown", 0)
-        self.restoreKeyDown = owner
-    end
-end
 
-function ActionButtons:RestoreActionButtonUseKeyDown(owner)
-    if self.restoreKeyDown == owner then
-        SetCVar("ActionButtonUseKeyDown", 1)
-        self.restoreKeyDown = nil
+function ActionButtons:SetShowSpellAnimations(enable)
+    local f = ActionBarActionEventsFrame
+
+    if enable then
+        if not f:IsEventRegistered("UNIT_SPELLCAST_SENT") then
+            f:RegisterEvent("UNIT_SPELLCAST_SENT")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_CHANNEL_START", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_CHANNEL_STOP", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_EMPOWER_START", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_EMPOWER_STOP", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_FAILED", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_INTERRUPTED", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_RETICLE_CLEAR", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_RETICLE_TARGET", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_START", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_STOP", "player")
+            f:RegisterUnitEvent("UNIT_SPELLCAST_SUCCEEDED", "player")
+        end
+    else
+        if f:IsEventRegistered("UNIT_SPELLCAST_SENT") then
+            f:UnregisterEvent("UNIT_SPELLCAST_SENT")
+            f:UnregisterEvent("UNIT_SPELLCAST_CHANNEL_START")
+            f:UnregisterEvent("UNIT_SPELLCAST_CHANNEL_STOP")
+            f:UnregisterEvent("UNIT_SPELLCAST_EMPOWER_START")
+            f:UnregisterEvent("UNIT_SPELLCAST_EMPOWER_STOP")
+            f:UnregisterEvent("UNIT_SPELLCAST_FAILED")
+            f:UnregisterEvent("UNIT_SPELLCAST_INTERRUPTED")
+            f:UnregisterEvent("UNIT_SPELLCAST_RETICLE_CLEAR")
+            f:UnregisterEvent("UNIT_SPELLCAST_RETICLE_TARGET")
+            f:UnregisterEvent("UNIT_SPELLCAST_START")
+            f:UnregisterEvent("UNIT_SPELLCAST_STOP")
+            f:UnregisterEvent("UNIT_SPELLCAST_SUCCEEDED")
+        end
     end
 end
 
@@ -578,56 +347,31 @@ end
 -- collection metamethods
 function ActionButtons:ForAll(method, ...)
     for button in pairs(self.buttons) do
-        local callback = button[method]
-        if type(callback) == "function" then
-            callback(button, ...)
-        else
-            error(("ActionButton %d does not have a method named %q"):format(button.id, method))
-        end
+        button[method](button, ...)
     end
 end
 
 function ActionButtons:ForAllWhere(predicate, method, ...)
-    for action, buttons in pairs(self.actionButtons) do
-        if next(buttons) ~= nil and predicate(action) then
-            for button in pairs(buttons) do
-                local callback = button[method]
-                if type(callback) == "function" then
-                    callback(button, ...)
-                else
-                    error(("ActionButton %d does not have a method named %q"):format(button.id, method))
-                end
-            end
+    for button, action in pairs(self.buttons) do
+        if predicate(action) then
+            button[method](button, ...)
         end
     end
 end
 
 function ActionButtons:ForActionSlot(slot, method, ...)
-    local actions = rawget(self.actionButtons, slot)
-
-    if actions ~= nil then
-        for button in pairs(actions) do
-            local callback = button[method]
-            if type(callback) == "function" then
-                callback(button, ...)
-            else
-                error(("ActionButton %d does not have a method named %q"):format(button.id, method))
-            end
+    for button, action in pairs(self.buttons) do
+        if action == slot then
+            button[method](button, ...)
         end
     end
 end
 
 function ActionButtons:ForSpellID(spellID, method, ...)
-    for action, buttons in pairs(self.actionButtons) do
-        if next(buttons) ~= nil and HasSpellID(action, spellID) then
-            for button in pairs(buttons) do
-                local callback = button[method]
-                if type(callback) == "function" then
-                    callback(button, ...)
-                else
-                    error(("ActionButton %d does not have a method named %q"):format(button.id, method))
-                end
-            end
+    local hasSpellID = HasSpellID
+    for button, action in pairs(self.buttons) do
+        if  hasSpellID(action, spellID) then
+            button[method](button, ...)
         end
     end
 end
@@ -635,12 +379,7 @@ end
 function ActionButtons:ForVisible(method, ...)
     for button in pairs(self.buttons) do
         if button:IsVisible() then
-            local callback = button[method]
-            if type(callback) == "function" then
-                callback(button, ...)
-            else
-                error(("ActionButton %d does not have a method named %q"):format(button.id, method))
-            end
+            button[method](button, ...)
         end
     end
 end
@@ -648,12 +387,7 @@ end
 function ActionButtons:ForVisibleWhere(predicate, method, ...)
     for button, action in pairs(self.buttons) do
         if button:IsVisible() and predicate(action) then
-            local callback = button[method]
-            if type(callback) == "function" then
-                callback(button, ...)
-            else
-                error(("ActionButton %d does not have a method named %q"):format(button.id, method))
-            end
+            button[method](button, ...)
         end
     end
 end